Extract position calculation
