+++ flatgfa/src/ops/mod.rs
@@ -1,1 +1,2 @@
 pub mod depth;
+pub mod position;
